requirements.txt
livekit-agents
livekit-plugins-silero
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-elevenlabs
python-dotenv
